Fix precondition in DcpHostService (#1943)
* Fix precondition in DcpHostService

Two lines are changed here, though it's the second (in `StopAsync`) that has a functional change.

The first guard clause allows a null publisher, or a "dcp" publisher.

The second clause would allow any non-null publisher, which looks like a bug to me. Specifically, the first half of the || excludes "dcp", then the second half allows it back in. This means that there's no point checking for "dcp" in the code as written.

My interpretation of this code is that both clauses are intended to be the same, and this commit brings them in line with one another.

* Deduplicate condition