Show OLTP timestamps using user's time format preferences (#2135)
We want to show timestamps with millisecond-level precision, however there's no built-in format string for that that respects the user's preferences. Microsoft docs provided the approach taken here, which uses the user's default pattern (which comes from the machines culture and may be overridden by the user at the system level), and modifies it to append milliseconds after the seconds. It respects the user's preference for decimal separators too.